+++ strack-app/include/Themes/LightTheme.h
