+++ strack-app/include/Themes/Theme.h
