+++ strack-app/include/Windows/MainWindow/Components/ApplicationMenuBar.h
