+++ strack-app/include/Windows/MainWindow/Components/CodeEditorWidget.h
