+++ strack-app/include/Windows/MainWindow/Components/ControlStripWidget.h
