+++ strack-app/include/Windows/MainWindow/Components/FileExplorerWidget.h
