+++ strack-app/include/Windows/MainWindow/MainWindow.h
