+++ strack-app/include/Windows/SettingsWindow/SettingsWindow.h
