+++ strack-app/include/Windows/StartupWindow/Components/ProjectActionsWidget.h
