+++ strack-app/include/Windows/StartupWindow/Components/RecentProjectsWidget.h
