+++ strack-app/include/Windows/StartupWindow/StartupWindow.h
